remove test/hits.test.js as no longer required #54
