+++ server.js
@@ -20,9 +20,9 @@ io.on('connection', function (socket) {
 });
 
 function handler (req, res) {
-  var url = req.url;
-  var hit = extract(req);
-  // console.log(hit);
+  var url = req.url;      // alias req.url to reduce typing in matching below
+  var hit = extract(req); // extract just the data we want from the HTTP request
+
   if (url.match(/svg/)) {      // only return a badge if SVG requested
     hits(hit, function(err, count) {
       io.sockets.emit('hit', { 'hit': format(hit, count) }); // broadcast
