+++ test/db_filesystem.test.js
@@ -26,6 +26,25 @@ test(file+'Add a hit to the list for that url', function (t) {
   })
 });
 
+test(file+'test with full github url', function (t) {
+  var hits = require('../lib/db_filesystem.js');
+  var req = {
+    'url': 'https://github.com/my/awesome/url',
+    headers: {
+      'accept-language': 'en-US,en;q=0.8,pt;q=0.6,es;',
+      'user-agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_5)'
+    },
+    connection: {
+      remoteAddress: '88.8.88.8'
+    }
+  }
+  var hit = extract(req);
+  hits(hit, function (err, count) {
+    t.ok(count >= 0, '✓ REQ ' +req.url +' was added at a index: ' + count);
+    t.end(); // shutdown redis con
+  })
+});
+
 test(file + 'Add a hit for new url', function (t) {
   var db = require('../lib/db_filesystem.js');
   var req = {
